% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsetup-package.R
\docType{package}
\name{pkgsetup-package}
\alias{pkgsetup}
\alias{pkgsetup-package}
\title{pkgsetup: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID}) [copyright holder]

}
\keyword{internal}
